Redirect to service page if is cookie and authenticated in public page
