+++ src/common/contants.ts
@@ -1,0 +1,1 @@
+export const SESSION_KEY = 'JSESSIONID';
