+++ src/hooks/index.ts
@@ -1,2 +1,2 @@
-export * from './useAuthorization';
 export * from './useDidMount';
+export * from './useDidMountEffect';
