+++ src/utils/error/isUnauthorizeError.ts
@@ -1,9 +1,9 @@
 import { AxiosError } from 'axios';
-import { Enum } from 'common';
+import { Enums } from 'common';
 
 export function isUnauthorizeError(error: unknown) {
   if (error instanceof AxiosError) {
-    if (error.response?.status === Enum.ErrorCode.UNAUTHORIZED) {
+    if (error.response?.status === Enums.ErrorCode.UNAUTHORIZED) {
       return true;
     }
   }
